+++ index.jsx
@@ -6,7 +6,7 @@ import App from "./src/App";
 const rootID = 'danmaku-migrate_root';
 const rootRef = document.createElement('DIV');
 rootRef.setAttribute('id', rootID);
-rootRef.classList = "absolute top-0 left-0 w-full overflow-visible z-20";
+rootRef.classList = "absolute top-0 left-0 w-full overflow-visible z-1001";
 document.body.appendChild(rootRef);
 
 setTimeout(() => {